requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
twilio==8.10.0
openai==1.3.7
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
